extract Client subscriptions into service
This change extracts the Client subscriptions logic into a separate
`startstop.Service` which can be started and stopped along with the
other services. The important change that enables this is switching from
a _callback_ for job events to a _channel_ for job events. The channel
is passed to the completer during init, and the completer then owns it
as the sole sender. When the completer is stopped, it must close the
channel to indicate that there are no more job completion events to be
processed.

This moves us closer to having all the key client services be able to
be managed as a single pool of services, and they can all have their
shutdown initiated in parallel. Importantly, this paves the way for
additional services to be added (even by external libraries) without
needing to deal with more complex startup & shutdown ordering scenarios.

In order to make this work with a client that can be started and stopped
repeatedly, a new `ResetSubscribeChan` method was added to the
`JobCompleter` interface to be called at the beginning of each
`Client.Start()` call.